Base class for objects using a pytorch model as core model.

Objects of this class containing fields and methods used in several other classes in 'AI for Education'.

This class is not designed for a direct application and should only be used by developers.

Value

A new object of this class.

See also

Super class

aifeducation::AIFEMaster -> AIFEBaseModel

Methods

Public methods

AIFEBaseModel$count_parameter()

AIFEBaseModel$clone()

Inherited methods

aifeducation::AIFEMaster$get_all_fields()

aifeducation::AIFEMaster$get_documentation_license()

aifeducation::AIFEMaster$get_ml_framework()

aifeducation::AIFEMaster$get_model_config()

aifeducation::AIFEMaster$get_model_description()

aifeducation::AIFEMaster$get_model_info()

aifeducation::AIFEMaster$get_model_license()

aifeducation::AIFEMaster$get_package_versions()

aifeducation::AIFEMaster$get_private()

aifeducation::AIFEMaster$get_publication_info()

aifeducation::AIFEMaster$get_sustainability_data()

aifeducation::AIFEMaster$is_configured()

aifeducation::AIFEMaster$is_trained()

aifeducation::AIFEMaster$set_documentation_license()

aifeducation::AIFEMaster$set_model_description()

aifeducation::AIFEMaster$set_model_license()

aifeducation::AIFEMaster$set_publication_info()

Method count_parameter()

Method for counting the trainable parameters of a model.

Usage

AIFEBaseModel$count_parameter()

Returns

Returns the number of trainable parameters of the model.

Method clone()

The objects of this class are cloneable with this method.

Usage

AIFEBaseModel$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.