Base class for most objects

Objects of this class containing fields and methods used in several other classes in 'AI for Education'.

This class is not designed for a direct application and should only be used by developers.

Value

A new object of this class.

See also

Public fields

last_training

('list()')
List for storing the history, the configuration, and the results of the last training. This information will be overwritten if a new training is started.

last_training$start_time: Time point when training started.

last_training$learning_time: Duration of the training process.

last_training$finish_time: Time when the last training finished.

last_training$history: History of the last training.

last_training$data: Object of class table storing the initial frequencies of the passed data.

last_training$config: List storing the configuration used for the last training.

Methods

Public methods

AIFEMaster$get_model_info()

AIFEMaster$set_publication_info()

AIFEMaster$get_publication_info()

AIFEMaster$set_model_license()

AIFEMaster$get_model_license()

AIFEMaster$set_documentation_license()

AIFEMaster$get_documentation_license()

AIFEMaster$set_model_description()

AIFEMaster$get_model_description()

AIFEMaster$get_package_versions()

AIFEMaster$get_sustainability_data()

AIFEMaster$get_ml_framework()

AIFEMaster$is_configured()

AIFEMaster$is_trained()

AIFEMaster$get_private()

AIFEMaster$get_all_fields()

AIFEMaster$get_model_config()

AIFEMaster$clone()

AIFEMaster$get_model_info()

Method for requesting the model information.

Usage

AIFEMaster$get_model_info()

Returns

list of all relevant model information.

AIFEMaster$set_publication_info()

Method for setting publication information of the model.

Usage

AIFEMaster$set_publication_info(authors, citation, url = NULL)

Arguments

authors

List of authors.

citation

Free text citation.

url

URL of a corresponding homepage.

Returns

Function does not return a value. It is used for setting the private members for publication information.

AIFEMaster$get_publication_info()

Method for requesting the bibliographic information of the model.

Usage

AIFEMaster$get_publication_info()

Returns

list with all saved bibliographic information.

AIFEMaster$set_model_license()

Method for setting the license of the model.

Usage

AIFEMaster$set_model_license(license = "CC BY")

Arguments

license

string containing the abbreviation of the license or the license text.

Returns

Function does not return a value. It is used for setting the private member for the software license of the model.

AIFEMaster$get_model_license()

Method for getting the license of the model.

Usage

AIFEMaster$get_model_license()

Returns

string representing the license for the model.

AIFEMaster$set_documentation_license()

Method for setting the license of the model's documentation.

Usage

AIFEMaster$set_documentation_license(license = "CC BY")

Arguments

license

string containing the abbreviation of the license or the license text.

Returns

Function does not return a value. It is used for setting the private member for the documentation license of the model.

AIFEMaster$get_documentation_license()

Method for getting the license of the model's documentation.

Usage

AIFEMaster$get_documentation_license()

Returns

Returns the license as a string.

AIFEMaster$set_model_description()

Method for setting a description of the model.

Usage

AIFEMaster$set_model_description(
  eng = NULL,
  native = NULL,
  abstract_eng = NULL,
  abstract_native = NULL,
  keywords_eng = NULL,
  keywords_native = NULL
)

Arguments

eng

string A text describing the training, its theoretical and empirical background, and output in English.

native

string A text describing the training , its theoretical and empirical background, and output in the native language of the model.

abstract_eng

string A text providing a summary of the description in English.

abstract_native

string A text providing a summary of the description in the native language of the model.

keywords_eng

vector of keyword in English.

keywords_native

vector of keyword in the native language of the model.

Returns

Function does not return a value. It is used for setting the private members for the description of the model.

AIFEMaster$get_model_description()

Method for requesting the model description.

Usage

AIFEMaster$get_model_description()

Returns

list with the description of the classifier in English and the native language.

AIFEMaster$get_package_versions()

Method for requesting a summary of the R and python packages' versions used for creating the model.

Usage

AIFEMaster$get_package_versions()

Returns

Returns a list containing the versions of the relevant R and python packages.

AIFEMaster$get_sustainability_data()

Method for requesting a summary of tracked energy consumption during training and an estimate of the resulting CO2 equivalents in kg.

Usage

AIFEMaster$get_sustainability_data(track_mode = "training")

Arguments

track_mode

string Determines the stept to which the data refer. Allowed values:

'training'

'inference'

Returns

Returns a list containing the tracked energy consumption, CO2 equivalents in kg, information on the tracker used, and technical information on the training infrastructure.

AIFEMaster$get_ml_framework()

Method for requesting the machine learning framework used for the model.

Usage

AIFEMaster$get_ml_framework()

Returns

Returns a string describing the machine learning framework used for the classifier.

AIFEMaster$is_configured()

Method for checking if the model was successfully configured. An object can only be used if this value is TRUE.

Usage

AIFEMaster$is_configured()

Returns

bool TRUE if the model is fully configured. FALSE if not.

AIFEMaster$is_trained()

Check if the TEFeatureExtractor is trained.

Usage

AIFEMaster$is_trained()

Returns

Returns TRUE if the object is trained and FALSE if not.

AIFEMaster$get_private()

Method for requesting all private fields and methods. Used for loading and updating an object.

Usage

AIFEMaster$get_private()

Returns

Returns a list with all private fields and methods.

AIFEMaster$get_all_fields()

Return all fields.

Usage

AIFEMaster$get_all_fields()

Returns

Method returns a list containing all public and private fields of the object.

AIFEMaster$get_model_config()

Method for requesting the model configuration.

Usage

AIFEMaster$get_model_config()

Returns

Returns a list with all configuration parameters used during configuration.

AIFEMaster$clone()

The objects of this class are cloneable with this method.

Usage

AIFEMaster$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.