Text embedding classifier with a ProtoNet

Abstract class for neural nets with 'keras'/'tensorflow' and 'pytorch'.

This object represents in implementation of a prototypical network for few-shot learning as described by Snell, Swersky, and Zemel (2017). The network uses a multi way contrastive loss described by Zhang et al. (2019). The network learns to scale the metric as described by Oreshkin, Rodriguez, and Lacoste (2018)

Value

Objects of this class are used for assigning texts to classes/categories. For the creation and training of a classifier an object of class EmbeddedText and a factor

are necessary. The object of class EmbeddedText contains the numerical text representations (text embeddings) of the raw texts generated by an object of class TextEmbeddingModel. The factor contains the classes/categories for every text. Missing values (unlabeled cases) are supported. For predictions an object of class EmbeddedText has to be used which was created with the same text embedding model as for training.

References

Oreshkin, B. N., Rodriguez, P. & Lacoste, A. (2018). TADAM: Task dependent adaptive metric for improved few-shot learning. https://doi.org/10.48550/arXiv.1805.10123

Snell, J., Swersky, K. & Zemel, R. S. (2017). Prototypical Networks for Few-shot Learning. https://doi.org/10.48550/arXiv.1703.05175

Zhang, X., Nie, J., Zong, L., Yu, H. & Liang, W. (2019). One Shot Learning with Margin. In Q. Yang, Z.-H. Zhou, Z. Gong, M.-L. Zhang & S.-J. Huang (Eds.), Lecture Notes in Computer Science. Advances in Knowledge Discovery and Data Mining (Vol. 11440, pp. 305–317). Springer International Publishing. https://doi.org/10.1007/978-3-030-16145-3_24

See also

Super class

aifeducation::TEClassifierRegular -> TEClassifierProtoNet

Methods

Public methods

TEClassifierProtoNet$new()

TEClassifierProtoNet$train()

TEClassifierProtoNet$embed()

TEClassifierProtoNet$clone()

Inherited methods

aifeducation::TEClassifierRegular$check_embedding_model()

aifeducation::TEClassifierRegular$extract_features()

aifeducation::TEClassifierRegular$get_documentation_license()

aifeducation::TEClassifierRegular$get_ml_framework()

aifeducation::TEClassifierRegular$get_model_description()

aifeducation::TEClassifierRegular$get_model_info()

aifeducation::TEClassifierRegular$get_package_versions()

aifeducation::TEClassifierRegular$get_publication_info()

aifeducation::TEClassifierRegular$get_software_license()

aifeducation::TEClassifierRegular$get_sustainability_data()

aifeducation::TEClassifierRegular$get_text_embedding_model()

aifeducation::TEClassifierRegular$load_model()

aifeducation::TEClassifierRegular$predict()

aifeducation::TEClassifierRegular$save_model()

aifeducation::TEClassifierRegular$set_documentation_license()

aifeducation::TEClassifierRegular$set_model_description()

aifeducation::TEClassifierRegular$set_publication_info()

aifeducation::TEClassifierRegular$set_software_license()

Method new()

Creating a new instance of this class.

Usage

TEClassifierProtoNet$new(
  ml_framework = aifeducation_config$get_framework(),
  name = NULL,
  label = NULL,
  text_embeddings = NULL,
  use_fe = TRUE,
  fe_features = 128,
  fe_method = "lstm",
  fe_noise_factor = 0.2,
  targets = NULL,
  hidden = c(128),
  rec = c(128),
  rec_type = "gru",
  rec_bidirectional = FALSE,
  embedding_dim = 3,
  self_attention_heads = 0,
  intermediate_size = NULL,
  attention_type = "fourier",
  add_pos_embedding = TRUE,
  rec_dropout = 0.1,
  repeat_encoder = 1,
  dense_dropout = 0.4,
  recurrent_dropout = 0.4,
  encoder_dropout = 0.1,
  optimizer = "adam"
)

Arguments

ml_framework

string Framework to use for training and inference. ml_framework="tensorflow" for 'tensorflow' and ml_framework="pytorch" for 'pytorch'

name

Character Name of the new classifier. Please refer to common name conventions. Free text can be used with parameter label.

label

Character Label for the new classifier. Here you can use free text.

text_embeddings

An object of classTextEmbeddingModel.

use_fe

bool If TRUE a feature extractor is applied in order to reduce the dimensionality of the text embeddings.

fe_features

int determining the number of dimensions to which the dimension of the text embedding should be reduced.

fe_method

string Method to use for the feature extraction. "lstm" for an extractor based on LSTM-layers or "dense" for dense layers.

fe_noise_factor

double between 0 and a value lower 1 indicating how much noise should be added for the training of the feature extractor.

targets

factor containing the target values of the classifier.

hidden

vector containing the number of neurons for each dense layer. The length of the vector determines the number of dense layers. If you want no dense layer, set this parameter to NULL.

rec

vector containing the number of neurons for each recurrent layer. The length of the vector determines the number of dense layers. If you want no dense layer, set this parameter to NULL.

rec_type

string Type of the recurrent layers. rec_type="gru" for Gated Recurrent Unit and rec_type="lstm" for Long Short-Term Memory.

rec_bidirectional

bool If TRUE a bidirectional version of the reccurent layers is used.

embedding_dim

Int determining the dimensionality of the embedding.

self_attention_heads

integer determining the number of attention heads for a self-attention layer. Only relevant if attention_type="multihead"

intermediate_size

int determining the size of the projection layer within a each transformer encoder.

attention_type

string Choose the relevant attention type. Possible values are "fourier" and multihead.

add_pos_embedding

bool TRUE if positional embedding should be used.

rec_dropout

double ranging between 0 and lower 1, determining the dropout between bidirectional gru layers.

repeat_encoder

int determining how many times the encoder should be added to the network.

dense_dropout

double ranging between 0 and lower 1, determining the dropout between dense layers.

recurrent_dropout

double ranging between 0 and lower 1, determining the recurrent dropout for each recurrent layer. Only relevant for keras models.

encoder_dropout

double ranging between 0 and lower 1, determining the dropout for the dense projection within the encoder layers.

optimizer

Object of class keras.optimizers.

Returns

Returns an object of class TextEmbeddingClassifierNeuralNet which is ready for training.

Method train()

Method for training a neural net.

Usage

TEClassifierProtoNet$train(
  data_embeddings,
  data_targets,
  data_folds = 5,
  data_val_size = 0.25,
  use_sc = TRUE,
  sc_method = "dbsmote",
  sc_min_k = 1,
  sc_max_k = 10,
  use_pl = TRUE,
  pl_max_steps = 3,
  pl_max = 1,
  pl_anchor = 1,
  pl_min = 0,
  sustain_track = TRUE,
  sustain_iso_code = NULL,
  sustain_region = NULL,
  sustain_interval = 15,
  epochs = 40,
  batch_size = 35,
  fe_epochs = 1000,
  fe_val_size = 0.25,
  Ns = 5,
  Nq = 3,
  loss_alpha,
  loss_margin,
  dir_checkpoint,
  trace = TRUE,
  keras_trace = 2,
  pytorch_trace = 1
)

Arguments

data_embeddings

Object of class TextEmbeddingModel.

data_targets

Factor containing the labels for cases stored in data_embeddings. Factor must be named and has to use the same names used in data_embeddings.

data_folds

int determining the number of cross-fold samples.

data_val_size

double between 0 and 1, indicating the proportion of cases of each class which should be used for the validation sample during the estimation of the baseline model. The remaining cases are part of the training data.

use_sc

bool TRUE if the estimation should integrate balanced synthetic cases. FALSE if not.

sc_method

vector containing the methods for generating synthetic cases via 'smotefamily'. Multiple methods can be passed. Currently sc_method=c("adas"), sc_method=c("smote") and sc_method=c("dbsmote") are possible.

sc_min_k

int determining the minimal number of k which is used for creating synthetic units.

sc_max_k

int determining the maximal number of k which is used for creating synthetic units.

use_pl

bool TRUE if the estimation should integrate balanced pseudo-labeling. FALSE if not.

pl_max_steps

int determining the maximum number of steps during pseudo-labeling.

pl_max

double between 0 and 1, setting the maximal level of confidence for considering a case for pseudo-labeling.

pl_anchor

double between 0 and 1 indicating the reference point for sorting the new cases of every label. See notes for more details.

pl_min

double between 0 and 1, setting the minimal level of confidence for considering a case for pseudo-labeling.

sustain_track

bool If TRUE energy consumption is tracked during training via the python library codecarbon.

sustain_iso_code

string ISO code (Alpha-3-Code) for the country. This variable must be set if sustainability should be tracked. A list can be found on Wikipedia: https://en.wikipedia.org/wiki/List_of_ISO_3166_country_codes.

sustain_region

Region within a country. Only available for USA and Canada See the documentation of codecarbon for more information. https://mlco2.github.io/codecarbon/parameters.html

sustain_interval

integer Interval in seconds for measuring power usage.

epochs

int Number of training epochs.

batch_size

int Size of the batches for the feature extractor.

fe_epochs

int Number of training epochs for the feature extractor.

fe_val_size

double between 0 and 1, indicating the proportion of cases which should be used for the validation sample during the training of the feature extractor.

Ns

int Number of cases for every class in the sample.

Nq

int Number of cases for every class in the query.

loss_alpha

double Value between 0 and 1 indicating how strong the loss should focus on pulling cases to its corresponding prototypes or pushing cases away from other prototypes. The higher the value the more the loss concentrates on pulling cases to its corresponding prototypes.

loss_margin

dobule Value greater 0 indicating the minimal distance of every case from prototypes of other classes.

dir_checkpoint

string Path to the directory where the checkpoint during training should be saved. If the directory does not exist, it is created.

trace

bool TRUE, if information about the estimation phase should be printed to the console.

keras_trace

int keras_trace=0 does not print any information about the training process from keras on the console.

pytorch_trace

int pytorch_trace=0 does not print any information about the training process from pytorch on the console. pytorch_trace=1 prints a progress bar.

balance_class_weights

bool If TRUE class weights are generated based on the frequencies of the training data with the method Inverse Class Frequency'. If FALSE each class has the weight 1.

balance_sequence_length

bool If TRUE sample weights are generated for the length of sequences based on the frequencies of the training data with the method Inverse Class Frequency'. If FALSE each sequences length has the weight 1.

Details

sc_max_k: All values from sc_min_k up to sc_max_k are successively used. If the number of sc_max_k is too high, the value is reduced to a number that allows the calculating of synthetic units.

pl_anchor: With the help of this value, the new cases are sorted. For this aim, the distance from the anchor is calculated and all cases are arranged into an ascending order.

Returns

Function does not return a value. It changes the object into a trained classifier.

Method embed()

Usage

TEClassifierProtoNet$embed(
  embeddings_q = NULL,
  classes_q = NULL,
  embeddings_s = NULL,
  classes_s = NULL
)

Method clone()

The objects of this class are cloneable with this method.

Usage

TEClassifierProtoNet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.