Text Embedding Classifier with a Neural Net

Abstract class for neural nets with keras and tensorflow

Public fields

name

('character()')
Name of the classifier.

label

('character()')
Label of the classifier used as the individual title.

date

('date()')
Date of the first creation of the classifier.

text_embedding_model

('list()')
List for storing information about the underlying text embedding model. This information ensures that a classifier is only used with data from the correct text embedding model and ensures the correct handling of the embeddings.

bundeled_model

('bundle object')
Object for storing the keras model of the neural net. Saved as bundled object with help of the package bundle for serialization.

model_config

('list()')
List for storing information about the configuration of the model. This information is used for predicting new data.

model_config$n_gru: Number of gru layers.

model_config$n_hidden: Number of dense layers.

model_config$target_levels: Levels of the target variable. Do not change this manually.

model_config$input_variables: Order and name of the input variables. Do not change this manually.

model_config$init_config: List storing all parameters passed to method new().

last_training

('list()')
List for storing the history and the results of the last training. This information will be overwritten if a new training is started.

last_training$learning_time: Duration of the training process.

config$history: History of the last training.

config$data: Object of class table storing the initial frequencies of the passed data.

config$data_pb:l Matrix storing the number of additional cases (test and training) added during balanced pseudo labeling. The rows refer to folds and final training. The columns refer to the steps during pseudo labeling.

config$data_bsc_test: Matrix storing the number of cases for each category used for testing during the phase of balanced synthetic units. Please note that the frequencies include original and synthetic cases. In terms that the number of original and synthetic cases exceeds the limit for the majority classes the frequency represents the number of cases created by cluster analysis.

config$date: Time when the last training finished.

config$config: List storing which kind of estimation was requested during last training.

config$config$use_bsc: TRUE if balanced synthetic cases was requested. FALSE if not.

config$config$use_baseline: TRUE if baseline estimation was requested. FALSE if not.

config$config$use_bpl: TRUE if balanced pseudo labeling cases was requested. FALSE if not.

reliability

('list()')
List for storing central reliability measures of the last training.

reliability$val_metric: Array containing the reliability measures for the validation data for every fold, method, and step (in case of pseudo labeling).

reliability$val_metric_mean: Array containing the reliability measures for the validation data for every method and step (in case of pseudo labeling). The values represent the mean values for every fold.

reliability$raw_iota_objects: List containing all iota_object generated with the package iotarelr for every fold at the start and the end of the last training.

reliability$raw_iota_objects$iota_objects_start: List of objects with class iotarelr_iota2 containing the estimated iota reliability of the second generation for the baseline model for every fold. If the estimation of the baseline model is not requested the list is set to NULL.

reliability$raw_iota_objects$iota_objects_end: List of objects with class iotarelr_iota2 containing the estimated iota reliability of the second generation for the final model for every fold. Depending of the requested training method these values refer to the baseline model, a trained model on the basis of balanced synthetic cases, balanced pseudo labeling or a combination of balanced synthetic cases with pseudo labeling.

reliability$raw_iota_objects$iota_objects_start_free: List of objects with class iotarelr_iota2 containing the estimated iota reliability of the second generation for the baseline model for every fold. If the estimation of the baseline model is not requested the list is set to NULL.Please note that the model is estimated without forcing the Assignment Error Matrix to be in line with the assumption of weak superiority.

reliability$raw_iota_objects$iota_objects_end_free: List of objects with class iotarelr_iota2 containing the estimated iota reliability of the second generation for the final model for every fold. Depending of the requested training method these values refer to the baseline model, a trained model on the basis of balanced synthetic cases, balanced pseudo labeling or a combination of balanced synthetic cases with pseudo labeling. Please note that the model is estimated without forcing the Assignment Error Matrix to be in line with the assumption of weak superiority.

reliability$iota_object_start: Object of class iotarelr_iota2 as a mean of the individual objects for every fold. If the estimation of the baseline model is not requested the list is set to NULL.

reliability$iota_object_start_free: Object of class iotarelr_iota2 as a mean of the individual objects for every fold. If the estimation of the baseline model is not requested the list is set to NULL. Please note that the model is estimated without forcing the Assignment Error Matrix to be in line with the assumption of weak superiority.

reliability$iota_object_end: Object of class iotarelr_iota2 as a mean of the individual objects for every fold. Depending of the requested training method this object refers to the baseline model, a trained model on the basis of balanced synthetic cases, balanced pseudo labeling or a combination of balanced synthetic cases with pseudo labeling.

reliability$iota_object_end_free: Object of class iotarelr_iota2 as a mean of the individual objects for every fold. Depending of the requested training method this object refers to the baseline model, a trained model on the basis of balanced synthetic cases, balanced pseudo labeling or a combination of balanced synthetic cases with pseudo labeling. Please note that the model is estimated without forcing the Assignment Error Matrix to be in line with the assumption of weak superiority.

Methods

Public methods

TextEmbeddingClassifierNeuralNet$new()

TextEmbeddingClassifierNeuralNet$train()

TextEmbeddingClassifierNeuralNet$predict()

TextEmbeddingClassifierNeuralNet$set_publication_info()

TextEmbeddingClassifierNeuralNet$get_publication_info()

TextEmbeddingClassifierNeuralNet$set_license()

TextEmbeddingClassifierNeuralNet$get_license()

TextEmbeddingClassifierNeuralNet$set_model_description()

TextEmbeddingClassifierNeuralNet$get_model_description()

TextEmbeddingClassifierNeuralNet$export_model()

TextEmbeddingClassifierNeuralNet$import_model()

TextEmbeddingClassifierNeuralNet$clone()

Method new()

Creating a new instance of this class.

Usage

TextEmbeddingClassifierNeuralNet$new(
  name = NULL,
  label = NULL,
  text_embeddings = NULL,
  targets = NULL,
  config = list(hidden = c(128), gru = c(128), dropout = 0.2, recurrent_dropout = 0,
    l2_regularizer = 0.001, optimizer = "adam", act_fct = "gelu", act_fct_last =
    "softmax", err_fct = "categorical_crossentropy")
)

Arguments

name

Character Name of the new classifier. Please refer to common name conventions. Free text can be used with parameter label.

label

Character Label for the new classifier. Here you can use free text.

text_embeddings

An object of classTextEmbeddingModel.

targets

factor containing the target values of the classifier.

config

list containing the configuration of the neural net.

config$hidden

vector containing the number of neurons for each dense layer. The length of the vector determines the number of dense layers. If you want no dense layer set this parameter to NULL.

config$gru

vector containing the number of neurons for each gru layer. The length of the vector determines the number of dense layers. If you want no dense layer set this parameter to NULL.

config$dropout

double ranging between 0 and lower 1 determining the dropout for each gru layer.

config$recurrent_dropout

double ranging between 0 and lower 1 determining the recurrent dropout for each gru layer.

l2_regularizer

double determining the l2 regularizer for every dense layer.

optimizer

Object of class keras.optimizers.

act_fct

character naming the activation function for all dense layers.

act_fct_last

character naming the activation function for the last dense layers.

err_fct

character naming the loss/error function for the neural net.

Method train()

Method for training a neural net with keras and tensorflow.

Usage

TextEmbeddingClassifierNeuralNet$train(
  data_embeddings,
  data_targets,
  data_n_valid_samples = 10,
  use_baseline = TRUE,
  bsl_val_size = 0.25,
  use_bsc = TRUE,
  bsc_methods = c("dbsmote"),
  bsc_max_k = 10,
  use_bpl = TRUE,
  bpl_max_steps = 10,
  bpl_inc_ratio = 0.25,
  bpl_anchor = 0.75,
  bpl_valid_size = 0.33,
  opt_model_reset = FALSE,
  epochs = 100,
  batch_size = 32,
  dir_checkpoint,
  trace = TRUE,
  view_metrics = FALSE,
  keras_trace = 2,
  n_cores = 2
)

Arguments

data_embeddings

Object of class TextEmbeddingModel.

data_targets

Factor containing the labels for cases stored in data_embeddings. Factor must be named and has to use the same names used in data_embeddings.

data_n_valid_samples

int determining the number of cross-fold samples.

use_baseline

bool TRUE if the calculation of a baseline model is requested. This option is only relevant for use_bsc=TRUE or use_pbl=TRUE. If both are FALSE a baseline model is calculated.

bsl_val_size

double between 0 and 1 indicating the proportion of cases of each label which should be used for the validation sample during the estimation of the baseline model. The remaining cases are part of the training data.

use_bsc

bool TRUE if the estimation should integrate balanced synthetic cases. FALSE if not.

bsc_methods

vector containing the methods for generating synthetic cases via smotefamily. Multiple methods can be passed. Currently bsc_methods=c("adas"), bsc_methods=c("smote") and bsc_methods=c("dbsmote") are possible.

bsc_max_k

int determining the maximal number of k which is used for creating synthetic units.

use_bpl

bool TRUE if the estimation should integrate balanced pseudo labeling. FALSE if not.

bpl_max_steps

int determining the maximal number of steps for every application of balanced pseudo labeling.

bpl_inc_ratio

double ratio between 0 and 1 indicating the proportion of new cases which should used for further training. See notes for more details.

bpl_anchor

double between 0 and 1 indicating the reference point for sorting the new cases of every label. See notes for more details.

bpl_valid_size

double ratio between 0 and 1 determining the proportion of new cases for every label which should be added to the validation sample during training. The remaining cases are added to the training sample.

opt_model_reset

bool TRUE if the model should be reseted before training.

epochs

int Number of training epochs.

batch_size

int Size of batches.

dir_checkpoint

string Path to the directory where the checkpoint during training should be saved. If the directory does not exists it is created.

trace

bool TRUE if information about the estimation phase should be printed to the console.

view_metrics

bool TRUE if metrics should be printed in the RStudie IDE.

keras_trace

int keras_trace=0 does not print any information about the training process from keras on the console. keras_trace=1 print a progress bar. keras_trace=2 prints one line of information for every epoch.

n_cores

int Number of cores used for creating synthetic units.

Details

bsc_max_k: All values from 2 up to bsc_max_k are successively used. If the number of bsc_max_k is to high the value is reduced to a number that allows the calculating of synthetic units.

bpl_inc_ratio: The ratio is applied to the label with the smallest number of new cases. The resulting value is used for every label to ensure the balance of all labels.

bpl_anchor: With the help of this value the new cases are sorted. For this aim the distance from the anchor is calculated and all cases are arranged into an increasing order.

Method predict()

Method for prediciting new data with a trained neural net.

Usage

TextEmbeddingClassifierNeuralNet$predict(newdata)

Arguments

newdata

Object of class TextEmbeddingModel or data.frame for which predictions should be made.

Returns

Returns a data.frame containing the predictions and the probabilities of the different labels for each case.

Method set_publication_info()

Method for setting publication information of the classifier

Usage

TextEmbeddingClassifierNeuralNet$set_publication_info(
  type,
  autors,
  citation,
  url = NULL
)

Arguments

type

string for choosing the type of information that should be added. type="developer" for information about the developer of the classifier and type="trainer" for information about persons who trained the classifier. If you modify an already existing classifier please use type="modifier".

autors

Person list of authors.

citation

Free text citation.

url

URL of a corresponding homepage.

Method get_publication_info()

Method for requesting the publication information of the classifier.

Usage

TextEmbeddingClassifierNeuralNet$get_publication_info()

Returns

list with all saved publication information.

Method set_license()

Method for setting the license of the classifier.

Usage

TextEmbeddingClassifierNeuralNet$set_license(license)

Arguments

license

string containing the abbreviation of the license or the license text.

Method get_license()

Method for requesting the license of the classifier.

Usage

TextEmbeddingClassifierNeuralNet$get_license()

Returns

string License of the classifier.

Method set_model_description()

Method for setting a description of the classifier.

Usage

TextEmbeddingClassifierNeuralNet$set_model_description(
  eng = NULL,
  native = NULL
)

Arguments

eng

string A text describing the training of the learner, its theoretical and empirical background, and the different output labels in English.

native

string A text describing the training of the learner, its theoretical and empirical background, and the different output labels in the native language of the classifier.

Method get_model_description()

Method for requesting the model description.

Usage

TextEmbeddingClassifierNeuralNet$get_model_description()

Returns

list with the description of the classifier in English and native language.

Method export_model()

Method for exporting a model to tensorflow SavedModel format.

Usage

TextEmbeddingClassifierNeuralNet$export_model(dir_path)

Arguments

dir_path

string() Path of the directory where the model should be saved.

Method import_model()

Method for importing a model from tensorflow SavedModel format.

Usage

TextEmbeddingClassifierNeuralNet$import_model(dir_path)

Arguments

dir_path

string() Path of the directory where the model is saved.

Method clone()

The objects of this class are cloneable with this method.

Usage

TextEmbeddingClassifierNeuralNet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.