Generate combinations of arguments

Function generates a specific number of combinations for a method. These are used for automating tests of objects.

Usage

generate_args_for_tests(
  object_name,
  method,
  var_objects = list(),
  necessary_objects = list(),
  var_override = list()
)

Arguments

object_name

string Name of the object to generate the arguments for.

method

string Name of the method of the object to generate the arguments for.

var_objects

list of other objects which should be combined with the other arguments.

necessary_objects

list of other objects which are part of every combination.

var_override

Named list containing the arguments which should be set to a specific value for all combinations.

Value

Returns a list with combinations of arguments.

Note

var_objects, necessary_objects, and var_override the names must exactly match the name of the parameter. Otherwise they are not applied. Names of arguments which are not part a a method are ignored. #'

See also