Get the number of chunks/sequences for each case

Function for calculating the number of chunks/sequences for every case.

Usage

get_n_chunks(text_embeddings, features, times, pad_value = -100L)

Arguments

text_embeddings

data.frame or array containing the text embeddings.

features

int Number of features within each sequence.

times

int Number of sequences.

pad_value

int Value indicating padding. This value should no be in the range of regluar values for computations. Thus it is not recommended to chance this value. Default is -100. Allowed values: \( x <= -1\)

Value

Namedvector of integers representing the number of chunks/sequences for every case.

See also