Install aifeducation on a machine

Function for installing 'aifeducation' on a machine.

Using a virtual environment (use_conda=FALSE) If 'python' is already installed the installed version is used. In the case that the required version of 'python' is different from the existing version the new version is installed. In all other cases python will be installed on the system.

#' Using a conda environment (use_conda=TRUE) If 'miniconda' is already existing on the machine no installation of 'miniconda' is applied. In this case the system checks for update and updates 'miniconda' to the newest version. If 'miniconda' is not found on the system it will be installed.

Usage

install_aifeducation(
  install_aifeducation_studio = TRUE,
  python_version = "3.12",
  cuda_version = "12.4",
  use_conda = FALSE
)

Arguments

install_aifeducation_studio

bool If TRUE all necessary R packages are installed for using AI for Education Studio.

python_version

string Python version to use/install.

cuda_version

string determining the requested version of cuda.

use_conda

bool If TRUE installation installs 'miniconda' and uses 'conda' as package manager. If FALSE installation installs python and uses virtual environments for package management.

Value

Function does nothing return. It installs python, optional R packages, and necessary 'python' packages on a machine.

Note

On MAC OS torch will be installed without support for cuda.

See also