Installing necessary python modules to an environment

Function for installing the necessary python modules.

Usage

install_py_modules(
  envname = "aifeducation",
  transformer_version = get_recommended_py_versions("transformers"),
  tokenizers_version = get_recommended_py_versions("tokenizers"),
  pandas_version = get_recommended_py_versions("pandas"),
  datasets_version = get_recommended_py_versions("datasets"),
  codecarbon_version = get_recommended_py_versions("codecarbon"),
  safetensors_version = get_recommended_py_versions("safetensors"),
  torcheval_version = get_recommended_py_versions("torcheval"),
  accelerate_version = get_recommended_py_versions("accelerate"),
  calflops_version = get_recommended_py_versions("calflops"),
  pytorch_cuda_version = "13.0",
  python_version = "3.12",
  remove_first = FALSE,
  use_conda = FALSE
)

Arguments

envname

string Name of the environment where the packages should be installed.

transformer_version

string determining the desired version of the python library 'transformers'.

tokenizers_version

string determining the desired version of the python library 'tokenizers'.

pandas_version

string determining the desired version of the python library 'pandas'.

datasets_version

string determining the desired version of the python library 'datasets'.

codecarbon_version

string determining the desired version of the python library 'codecarbon'.

safetensors_version

string determining the desired version of the python library 'safetensors'.

torcheval_version

string determining the desired version of the python library 'torcheval'.

accelerate_version

string determining the desired version of the python library 'accelerate'.

calflops_version

string determining the desired version of the python library 'calflops'.

pytorch_cuda_version

string determining the desired version of 'cuda' for 'PyTorch'. To install 'PyTorch' without cuda set to NULL.

python_version

string Python version to use.

remove_first

bool If TRUE removes the environment completely before recreating the environment and installing the packages. If FALSE the packages are installed in the existing environment without any prior changes.

use_conda

bool If TRUE uses 'conda' for package management. If FALSE uses virtual environments for package management.

Value

Returns no values or objects. Function is used for installing the necessary python libraries in a conda environment.

Note

Function tries to identify the type of operating system. In the case that MAC OS is detected 'PyTorch' is installed without support for cuda.

Supported versions of the packages can be requested with get_recommended_py_versions())

See also