Loading models created with 'aifeducation'

Function for loading models created with 'aifeducation'.

Usage

load_ai_model(model_dir, ml_framework = aifeducation_config$get_framework())

Arguments

model_dir

Path to the directory where the model is stored.

ml_framework

string Determines the machine learning framework for using the model. Possible are ml_framework="pytorch" for 'pytorch', ml_framework="tensorflow" for 'tensorflow', and ml_framework="auto". for using the framework used when saving the model.

Value

Returns an object of class TextEmbeddingClassifierNeuralNet or TextEmbeddingModel.

See also