Load target data for long running tasks

Function loads the target data for a long running task.

Usage

long_load_target_data(file_path, selectet_column)

Arguments

file_path

string Path to the file storing the target data.

selectet_column

string Name of the column containing the target data.

Value

Returns a named factor containing the target data.

Details

This function assumes that the target data is stored as a columns with the cases in the rows and the categories in the columns. The ids of the cases must be stored in a column called "id".

See also