Print message

Prints a message msg if trace parameter is TRUE with current date with message() or cat() function.

Usage

output_message(msg, trace, msg_fun)

Arguments

msg

string Message that should be printed.

trace

bool Silent printing (FALSE) or not (TRUE).

msg_fun

bool value that determines what function should be used. TRUE for message(), FALSE for cat().

Value

This function returns nothing.

See also