Saving objects created with 'aifeducation'

Function for saving objects created with 'aifeducation'.

Usage

save_to_disk(object, dir_path, folder_name)

Arguments

object

Object of class TEClassifierRegular, TEClassifierProtoNet, TEFeatureExtractor, TextEmbeddingModel, LargeDataSetForTextEmbeddings, LargeDataSetForText or EmbeddedText which should be saved.

dir_path

string Path to the directory where the should model is stored.

folder_name

string Name of the folder where the files should be stored.

Value

Function does not return a value. It saves the model to disk.

No return value, called for side effects.

See also