Summarize arguments from shiny input

This function extracts the input relevant for a specific method of a specific class from shiny input.

In addition, it adds the path to all objects which can not be exported to another R session. These object must be loaded separately in the new session with the function add_missing_args. The paths are intended to be used with shiny::ExtendedTask. The final preparation of the arguments should be done with

The function can also be used to override the default value of a method or to add value for arguments which are not part of shiny input (use parameter override_args).

Usage

summarize_args_for_long_task(
  input,
  object_class,
  method = "configure",
  path_args = list(path_to_embeddings = NULL, path_to_textual_dataset = NULL,
    path_to_target_data = NULL, path_to_feature_extractor = NULL, destination_path =
    NULL, folder_name = NULL),
  override_args = list(),
  meta_args = list(py_environment_type = get_py_env_type(), py_env_name =
    get_py_env_name(), target_data_column = input$data_target_column, object_class =
    input$classifier_type)
)

Arguments

input

Shiny input.

object_class

string Class of the object.

method

string Method of the class for which the arguments should be extracted and prepared.

path_args

list List containing the path to object that can not be exported to another R session. These must be loaded in the session.

override_args

list List containing all arguments that should be set manually. The values override default values of the argument and values which are part of input.

meta_args

list List containing information that are not relevant for the arguments of the method but are necessary to set up the shiny::ExtendedTask correctly.

Value

Returns a named list with the following entries:

args: Named list of all arguments necessary for the method of the class.

path_args: Named list of all paths for loading the objects missing in args.

meta_args: Named list of all arguments that are not part of the arguments of the method but which are necessary to set up the shiny::ExtendedTask correctly.

Note

Please not that all list are named list of the format (argument_name=values).

See also