Function for training and fine-tuning a bert model

This function can be used for training or fine-tuning a transformer based on Bert architecture with help of the python libraries 'transformers', 'datasets', and 'tokenizers'.

train_tune_bert_model(
  output_dir,
  bert_model_dir_path,
  raw_texts,
  aug_vocab_by = 100,
  p_mask = 0.15,
  whole_word = TRUE,
  val_size = 0.1,
  n_epoch = 1,
  batch_size = 12,
  chunk_size = 250,
  n_workers = 1,
  multi_process = FALSE,
  trace = TRUE
)

Arguments

output_dir

string Path to the directory where the final model should be saved. If the directory does not exists it will be created.

bert_model_dir_path

string Path to the directory where the original model is stored.

raw_texts

vector containing the raw texts for training.

aug_vocab_by

int Number of entries for extending the current vocabulary. See notes for more details

p_mask

double Ratio determining the number of words/tokens for masking.

whole_word

bool TRUE if whole word masking should be applied. If FALSE token masking is used.

val_size

double Ratio determining the amount of token chunks used for validation.

n_epoch

int Number of epochs for training.

batch_size

int Size of batches.

chunk_size

int Size of every chunk for training.

n_workers

int Number of workers.

multi_process

bool TRUE if multiple process should be activated.

trace

bool TRUE if information on the progress should be printed to the console.

Value

This function does not return an object. Instead the trained or fine-tuned model is saved to disk.

Note

if aug_vocab_by > 0 the raw texts is used for training a WordPiece tokenizer. At the end of this process additional entries are added to the vocabulary that are not part of the original vocabulary.

Pre-Trained models which can be fine-tuned with this function are available at https://huggingface.co/. New models can be created via the function create_bert_model.