Updates an existing installation of 'aifeducation' on a machine

Function for updating 'aifeducation' on a machine.

The function tries to find an existing environment on the machine, removes the environment and installs the environment with the new python modules.

In the case env_type = "auto" the function tries to update an existing virtual environment. If no virtual environment exits it tries to update a conda environment.

Usage

update_aifeducation(
  update_aifeducation_studio = TRUE,
  env_type = "auto",
  cuda_version = "12.4",
  envname = "aifeducation"
)

Arguments

update_aifeducation_studio

bool If TRUE all necessary R packages are installed for using AI for Education Studio.

env_type

string If set to "venv" virtual environment is requested. If set to "conda" a 'conda' environment is requested. If set to "auto" the function tries to use a virtual environment with the given name. If this environment does not exist it tries to activate a conda environment with the given name. If this fails the default virtual environment is used.'

cuda_version

string determining the requested version of cuda.

envname

string Name of the environment where the packages should be installed.

Value

Function does nothing return. It installs python, optional R packages, and necessary 'python' packages on a machine.

Note

On MAC OS torch will be installed without support for cuda.

See also