Vignette classifier trained with Synthetic Cases and Pseudo Labeling

An object of class TEClassifierProtoNet trained with the a subset of the Standford Movie Review Dataset. The purpose of classifier is for illustration in vignettes.

Usage

vignette_classifier_sc_pl

Format

R6