Embedded Text

Object of class R6 which stores the text embeddings generated by an object of class TextEmbeddingModel via the method embed().

Public fields

embeddings

('data.frame()')
data.frame containing the text embeddings for all chunks. Documents are in the rows. Embeddings dimensions are in the columns.

Methods

Public methods

EmbeddedText$new()

EmbeddedText$get_model_info()

EmbeddedText$get_model_label()

EmbeddedText$clone()

Method new()

Creates a new object representing text embeddings.

Usage

EmbeddedText$new(
  model_name,
  model_label,
  model_date,
  model_method,
  model_version,
  model_language,
  param_seq_length,
  param_chunks = NULL,
  param_overlap = NULL,
  param_aggregation = NULL,
  embeddings
)

Arguments

model_name

string Name of the model that generates this embedding.

model_label

string Label of the model that generates this embedding.

model_date

string Date when the embedding generating model was created.

model_method

string Method of the underlying embedding model.

model_version

string Version of the model that generated this embedding.

model_language

string Language of the model that generated this embedding.

param_seq_length

int Maximal number of tokens that processes the generating model for a chunk.

param_chunks

int Maximal number of chunks which are supported by the generating model.

param_overlap

int Number of tokens that were added at the beginning of the sequence for the next chunk by this model.

param_aggregation

string Aggregation method of the hidden states.

embeddings

data.frame containing the text embeddings.

Method get_model_info()

Method for retrieving information about the model that generated this embedding.

Usage

EmbeddedText$get_model_info()

Returns

list contain all saved information about the underlying text embedding model.

Method get_model_label()

Method for retrieving the label of the model that generated this embedding.

Usage

EmbeddedText$get_model_label()

Returns

string Lable of the corresponding text embedding model

Method clone()

The objects of this class are cloneable with this method.

Usage

EmbeddedText$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.