Question-Answer-Models of Type Extraction

This R6 class stores the information for modeling a question-answer-model. This kind of model extracts the answer from a given text.

Public fields

model

('transformers.TFAutoModelForQuestionAnswering')
Object of class transformers.TFAutoModelForQuestionAnswering from transformers python library. Stores the qa model.

tokenizer

('transformers.AutoTokenizer')
Object of class transformers.AutoTokenizer from transformers python library. Stores the tokenizer.

qa_pipline

('transformers.QuestionAnsweringPipeline')
Object of class transformers.QuestionAnsweringPipeline from transformers python library.

Methods

Public methods

QAExtractModel$new()

QAExtractModel$save_model()

QAExtractModel$load_model()

QAExtractModel$answer_question()

QAExtractModel$clone()

Method new()

Method for creating a new question-answer-model based on a pretrained model.

Usage

QAExtractModel$new(
  model_name,
  model_version,
  model_language,
  model_license,
  model_dir_path
)

Arguments

model_name

Character Name of the new model.

model_version

Character Version of the model.

model_language

Character Language the model supports.

model_license

Character License of the model.

model_dir_path

string Path to the directory where the model is stored.

Method save_model()

Method for saving a question-answer-model.

Usage

QAExtractModel$save_model(model_dir_path)

Arguments

model_dir_path

string Path to the directory where the model and the tokenizer should be stored.

Method load_model()

Method for loading a question-answer-model.

Usage

QAExtractModel$load_model(model_dir_path)

Arguments

model_dir_path

string Path to the directory where the model and the tokenizer are saved.

Method answer_question()

Method for extracting answers from a given text.

Usage

QAExtractModel$answer_question(
  question,
  knowledge_base,
  n_answers = 1,
  doc_stride = 128,
  max_answer_len = 15,
  max_seq_len = 384,
  max_question_len = 64,
  handle_impossible_answer = FALSE,
  align_to_words = TRUE
)

Arguments

question

string Question to be answered.

knowledge_base

list of raw texts where to search for the answer.

n_answers

int Number of possible answers generated from the texts.

doc_stride

int In case the knowledge base is too long for the model to process at once, the text is divided into several overlapping chunks. This parameter determines the size of the overlap.

max_answer_len

int Maximum length in tokens for possible answers. Only answers which are shorter are considered for an answer.

max_seq_len

int Maximum length of a question and knowledge base after tokenization. The context may be divided into several overlapping chunks.

max_question_len

int The maximum length of the question after tokenization. Longer sequences are truncated.

handle_impossible_answer

bool TRUE if impossible answers should be accepted.

align_to_words

bool If true TRUE the algorithm tries to align the answer to real words which increases the quality of the results for space separated languages.

Method clone()

The objects of this class are cloneable with this method.

Usage

QAExtractModel$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.