Text embedding model

This R6 class stores a text embedding model which can be used to tokenize, encode, decode, and embed raw texts. The object provides a unique interface for different text processing methods.

Public fields

basic_components

('list()')
List storing information which can apply to all methods.

basic_components$method: Method underlying the text embedding model.

basic_components$max_length: Maximum number of tokens in the sequence the model processes. In general, shorter sequences will be padded and longer sequences will be divided into chunks and/or truncated.

transformer_components

('list()')
List storing information which only apply to BERT models.

transformer_components$model: An object of class transformers.TFBertModel for using with transformers library.

transformer_components$tokenizer: An object of class transformers.BertTokenizerFast for using with transformers library.

transformer_components$aggregation: Aggregation method for the hidden states.

transformer_components$chunks: Maximal number of chunks processed with the model.

transformer_components$overlap: Number of tokens which should be added at the beginning of the sequence of the next chunk.

bow_components

('list()')
List storing information which only apply to bow_models.

bow_components$model: data.frame describing the relationship between tokens and their corresponding text embeddings.

bow_components$vocab: data.frame saving tokens, lemmas and their corresponding integer index.

bow_components$configuration: List of the configuration parameters of the model.

bow_components$configuration$to_lower TRUE if tokens are transformed to lower case.

bow_components$configuration$use_lemmata TRUE if the corresponding lemma should be used instead of the token.

bow_components$configuration$bow_n_dim Number of dimensions for GlobalVectors and Topic Modeling.

bow_components$configuration$bow_n_cluster Number of clusters for grouping tokens based in their global vectors. Does not apply to method lda.

bow_components$configuration$bow_max_iter Maximum number of iterations to calculate global vectors and topics.

bow_components$configuration$bow_max_iter_cluster Maximum number of iterations for creating clusters. Applies only to method glove.

bow_components$configuration$bow_cr_criterion Convergence criterion for calculating global vectors and topics.

bow_components$configuration$bow_learning_rate Initial learning rate for estimating global vectors.

bow_components$aggregation: Does currently not apply to these methods.

bow_components$chunks: Does currently not apply to these methods.

bow_components$overlap: Does currently not apply to these methods.

Methods

Public methods

TextEmbeddingModel$new()

TextEmbeddingModel$load_model()

TextEmbeddingModel$encode()

TextEmbeddingModel$decode()

TextEmbeddingModel$embed()

TextEmbeddingModel$set_publication_info()

TextEmbeddingModel$get_publication_info()

TextEmbeddingModel$set_license()

TextEmbeddingModel$get_license()

TextEmbeddingModel$set_model_description()

TextEmbeddingModel$get_model_description()

TextEmbeddingModel$get_model_info()

TextEmbeddingModel$clone()

Method new()

Method for creating a new text embedding model

Usage

TextEmbeddingModel$new(
  model_name = NULL,
  model_label = NULL,
  model_version = NULL,
  model_language = NULL,
  method = NULL,
  max_length = 0,
  chunks = 1,
  overlap = 0,
  aggregation = "last",
  model_dir,
  bow_basic_text_rep,
  bow_n_dim = 10,
  bow_n_cluster = 100,
  bow_max_iter = 500,
  bow_max_iter_cluster = 500,
  bow_cr_criterion = 1e-08,
  bow_learning_rate = 1e-08,
  trace = FALSE
)

Arguments

model_name

string containing the name of the new model.

model_label

string containing the label/title of the new model.

model_version

string version of the model.

model_language

string containing the language which the model represents (e.g., English).

method

string determining the kind of embedding model. Currently three types are supported. method="bert" for Bidirectional Encoder Representations from Transformers (BERT), method="glove" for GlobalVector Clusters, and method="lda" for topic modeling. See details for more information.

max_length

int determining the maximum length of token sequences uses in BERT models. Not relevant for the other methods.

chunks

int Maximum number of chunks. Only relevant for BERT models.

overlap

int determining the number of tokens which should be added at the beginning of the next chunk. Only relevant for BERT models.

aggregation

string method for aggregating the text embeddings created by BERT models. See details for more information.

model_dir

string path to the directory where the BERT model is stored.

bow_basic_text_rep

object of class basic_text_rep created via the function bow_pp_create_basic_text_rep. Only relevant for method="glove" and method="lda".

bow_n_dim

int Number of dimensions of the GlobalVector or number of topics for LDA.

bow_n_cluster

int Number of clusters created on the basis of GlobalVectors. Parameter is not relevant for method="lda" and method="bert"

bow_max_iter

int Maximum number of iterations for fitting GlobalVectors and Topic Models.

bow_max_iter_cluster

int Maximum number of iterations for fitting cluster if method="glove".

bow_cr_criterion

double convergence criterion for GlobalVectors.

bow_learning_rate

double initial learning rate for GlobalVectors.

trace

bool TRUE prints information about the progress. FALSE does not.

Details

method: In the case of method="bert", a pretrained BERT model must be supplied via model_dir. For method="glove" and method="lda" a new model will be created based on the data provided via bow_basic_text_rep. The original algorithm for GlobalVectors provides only word embeddings, not text embeddings. To achieve text embeddings the words are clustered based on their word embeddings with kmeans.

aggregation: For creating a text embedding with a BERT model, several options are possible:

last: aggregation="last" uses only the hidden states of the last layer.

second_to_last: aggregation="second_to_last" uses the hidden states of the second to last layer.

fourth_to_last: aggregation="fourth_to_last" uses the hidden states of the fourth to last layer.

all: aggregation="all" uses the mean of the hidden states of all hidden layers.

last_four: aggregation="last_four" uses the mean of the hidden states of the last four layers.

Method load_model()

Method for loading a transformers model into R.

Usage

TextEmbeddingModel$load_model(model_dir)

Arguments

model_dir

string containing the path to the relevant model.

Method encode()

Method for encoding words of raw texts into integers.

Usage

TextEmbeddingModel$encode(
  raw_text,
  token_encodings_only = FALSE,
  trace = FALSE
)

Arguments

raw_text

vector containing the raw texts.

token_encodings_only

bool If TRUE, only the token encodings are returned. If FALSE, the complete encoding is returned which is important for BERT models.

trace

bool If TRUE, information of the progress is printed. FALSE if not requested.

Returns

list containing the integer sequences of the raw texts with special tokens.

Method decode()

Method for decoding a sequence of integers into tokens

Usage

TextEmbeddingModel$decode(int_seqence)

Arguments

int_seqence

list containing the integer sequences which should be transformed to tokens or a single integer sequence as vector

Returns

list of token sequences

Method embed()

Method for creating text embeddings from raw texts

Usage

TextEmbeddingModel$embed(
  raw_text = NULL,
  doc_id = NULL,
  batch_size = 8,
  trace = FALSE
)

Arguments

raw_text

vector containing the raw texts.

doc_id

vector containing the corresponding IDs for every text.

batch_size

int determining the maximal size of every batch.

trace

bool TRUE, if information about the progression should be printed on console.

Returns

Method returns a R6 object of class EmbeddedText. This object contains the embeddings as a data.frame and information about the model creating the embeddings.

Method set_publication_info()

Method for setting the bibliographic information of the model.

Usage

TextEmbeddingModel$set_publication_info(type, authors, citation, url = NULL)

Arguments

type

string Type of information which should be changed/added. type="developer", and type="modifier" are possible.

authors

List of people.

citation

string Citation in free text.

url

string Corresponding URL if applicable.

Method get_publication_info()

Method for getting the bibliographic information of the model.

Usage

TextEmbeddingModel$get_publication_info()

Returns

list of bibliographic information.

Method set_license()

Method for setting the license of the model

Usage

TextEmbeddingModel$set_license(license)

Arguments

license

string containing the abbreviation of the license or the license text.

Method get_license()

Method for requesting the license of the model

Usage

TextEmbeddingModel$get_license()

Returns

string License of the model

Method set_model_description()

Method for setting a description of the model

Usage

TextEmbeddingModel$set_model_description(
  eng = NULL,
  native = NULL,
  abstract_eng = NULL,
  abstract_native = NULL,
  keywords_eng = NULL,
  keywords_native = NULL
)

Arguments

eng

string A text describing the training of the classifier, its theoretical and empirical background, and the different output labels in English.

native

string A text describing the training of the classifier, its theoretical and empirical background, and the different output labels in the native language of the model.

abstract_eng

string A text providing a summary of the description in English.

abstract_native

string A text providing a summary of the description in the native language of the classifier.

keywords_eng

vector of keywords in English.

keywords_native

vector of keywords in the native language of the classifier.

Method get_model_description()

Method for requesting the model description.

Usage

TextEmbeddingModel$get_model_description()

Returns

list with the description of the model in English and the native language.

Method get_model_info()

Method for requesting the model information

Usage

TextEmbeddingModel$get_model_info()

Returns

list of all relevant model information

Method clone()

The objects of this class are cloneable with this method.

Usage

TextEmbeddingModel$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.