Calculates reliability measures based on content analysis

This functions calculates different reliability measures which are based on the empirical research method of content analysis.

Usage

get_coder_metrics(true_values, predicted_values)

Arguments

true_values

factor containing the true labels/categories.

predicted_values

factor containing the predicted labels/categories.

Value

Returns a vector with the following reliability measures: #'

iota_index: Iota Index from the Iota Reliability Concept Version 2.

min_iota2: Minimal Iota from Iota Reliability Concept Version 2.

avg_iota2: Average Iota from Iota Reliability Concept Version 2.

max_iota2: Maximum Iota from Iota Reliability Concept Version 2.

min_alpha: Minmal Alpha Reliability from Iota Reliability Concept Version 2.

avg_alpha: Average Alpha Reliability from Iota Reliability Concept Version 2.

max_alpha: Maximum Alpha Reliability from Iota Reliability Concept Version 2.

static_iota_index: Static Iota Index from Iota Reliability Concept Version 2.

dynamic_iota_index: Dynamic Iota Index Iota Reliability Concept Version 2.

kalpha_nominal: Krippendorff's Alpha for nominal variables.

kalpha_ordinal: Krippendorff's Alpha for ordinal variables.

kendall: Kendall's coefficient of concordance W.

kappa2: Cohen's Kappa with equal weights.

kappa_fleiss: Fleiss' Kappa for multiple raters with exact estimation.

kappa_light: Light's Kappa for multiple raters.

percentage_agreement: Percentage Agreement.

gwet_ac: Gwet's AC1/AC2 agreement coefficient.