Get dictionary of all parameters

Function provides a list containing important characteristics of the parameter used in the models. The list does contain only the definition of arguments for transformer models and all classifiers. The arguments of other functions in this package are documented separately.

The aim of this list is to automatize argument checking and widget generation for AI for Education - Studio.

Usage

get_param_dict()

Value

Returns a named list. The names correspond to specific arguments. The list contains a list for every argument with the following components:

type: The type of allowed values.

allow_null: A bool indicating if the argument can be set to NULL.

min: The minimal value the argument can be. Set to NULL if not relevant. Set to -Inf if there is no minimum.

max: The maximal value the argument can be. Set to NULL if not relevant. Set to Inf if there is no Minimum.

desc: A string which includes the description of the argument written in markdown. This string is for the documentation the parameter.

values_desc: A named list containing a description of every possible value. The names must exactly match the strings in allowed_values. Descriptions should be written in markdown.

allowed_values: vector of allowed values. This is only relevant if the argument is not numeric. During the checking of the arguments it is checked if the provided values can be found in this vector. If all values are allowed set to NULL.

default_value: The default value of the argument. If there is no default set to NULL.

default_historic: Historic default value. This can be necessary for backward compatibility.

gui_box: string Name of the box in AI for Education - Studio where the argument appears. If it should not appear set to NULL.

gui_label: string Label of the controlling widget in AI for Education - Studio.

See also